requirements.txt
streamlit
openai
audio_recorder_streamlit
streamlit-float
transformers
streamlit_float
python-dotenv
langdetect
groq
pydub